pyproject.toml
[tool.poetry]
name = "invoice-processor"
version = "0.1.0"
description = "Automated invoice extraction and processing using OCR and AI"
authors = ["AIQSO <contact@aiqso.io>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/qvidal01/invoice-processor"
repository = "https://github.com/qvidal01/invoice-processor"
keywords = ["invoice", "ocr", "ai", "accounting", "automation", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Accounting",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = "^3.9"

# OCR & Document Processing
pytesseract = "^0.3.10"
pdf2image = "^1.16.3"
PyPDF2 = "^3.0.0"
Pillow = "^10.0.0"

# AI & NLP
openai = "^1.3.0"
langchain = "^0.1.0"
tiktoken = "^0.5.0"

# Data Validation & Processing
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"
python-dateutil = "^2.8.2"

# Integrations
requests = "^2.31.0"
python-quickbooks = "^0.9.0"
pyxero = "^0.9.0"
authlib = "^1.2.0"

# Storage & Queue
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
redis = {version = "^5.0.0", optional = true}

# Web Framework (for API/MCP server)
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}

# MCP Server
mcp = "^0.1.0"

# CLI
click = "^8.1.0"
rich = "^13.6.0"

# Utilities
python-dotenv = "^1.0.0"
cryptography = "^41.0.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.12.0"
faker = "^20.0.0"

# Code Quality
black = "^23.10.0"
ruff = "^0.1.0"
mypy = "^1.6.0"
pre-commit = "^3.5.0"

# Security
detect-secrets = "^1.4.0"

# Documentation
sphinx = "^7.2.0"
sphinx-rtd-theme = "^1.3.0"

# Development Tools
ipython = "^8.17.0"
ipdb = "^0.13.13"

[tool.poetry.extras]
redis = ["redis"]
all = ["redis"]

[tool.poetry.scripts]
invoice-processor = "invoice_processor.cli:main"
invoice-mcp-server = "invoice_processor.mcp_server.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# Mypy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = [
    "pytesseract.*",
    "pdf2image.*",
    "PyPDF2.*",
    "langchain.*",
    "quickbooks.*",
    "pyxero.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=invoice_processor --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
    "external: Tests that require external services",
]

# Coverage configuration
[tool.coverage.run]
source = ["invoice_processor"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
